:pencil: move place database package
