+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/AppDaoAccess.kt
@@ -3,6 +3,7 @@ package com.yoesuv.infomalangbatu.databases
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import com.yoesuv.infomalangbatu.databases.place.PlaceRoom
 
 @Dao
 interface AppDaoAccess {
