+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/place/DatabaseDeleteAllPlace.kt
@@ -1,7 +1,6 @@
 package com.yoesuv.infomalangbatu.databases.place
 
 import android.os.AsyncTask
-import com.yoesuv.infomalangbatu.databases.PlaceDatabase
 
 class DatabaseDeleteAllPlace(private val placeDatabase: PlaceDatabase) : AsyncTask<Void, Void, Void>() {
