+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/place/DatabaseInsertPlace.kt
@@ -1,8 +1,6 @@
 package com.yoesuv.infomalangbatu.databases.place
 
 import android.os.AsyncTask
-import com.yoesuv.infomalangbatu.databases.PlaceDatabase
-import com.yoesuv.infomalangbatu.databases.PlaceRoom
 
 class DatabaseInsertPlace(private val placeDatabase: PlaceDatabase, private val placeRoom: PlaceRoom): AsyncTask<Void, Void, Void>() {
