+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/place/PlaceDatabase.kt
@@ -1,7 +1,8 @@
-package com.yoesuv.infomalangbatu.databases
+package com.yoesuv.infomalangbatu.databases.place
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.yoesuv.infomalangbatu.databases.AppDaoAccess
 
 @Database(entities = [(PlaceRoom::class)], version =1)
 abstract class PlaceDatabase : RoomDatabase() {
